add : prompt for deleting

+++ backend/src/controllers/MedicineController.ts
@@ -1,5 +1,7 @@
 import _ from 'lodash';
 import Medicine from '../models/Medicine';
+import { database as sequelize } from '../configs/sqlite';
+import Prescription from '../models/Prescription';
 
 const MedicineController: IMedicineController = {
     async post(
@@ -52,23 +54,45 @@ const MedicineController: IMedicineController = {
     async delete(event: any, args: { id: number }) {
         try {
             const medicineId = args.id;
-            let deleteEntry = await Medicine.destroy({
-                where: {
-                    id: medicineId
-                }
-            });
+            const transaction_t1 = await sequelize.transaction();
 
-            if (deleteEntry === 1) {
+            if (args.id) {
+                // @ts-ignore
+                let deletePrescriptionWithCurrentMedicine =
+                    await Prescription.destroy({
+                        where: {
+                            medicine_id: args.id
+                        },
+                        transaction: transaction_t1
+                    });
+                let deleteEntry = await Medicine.destroy({
+                    where: {
+                        id: medicineId
+                    },
+
+                    transaction: transaction_t1
+                });
+
+                await transaction_t1.commit();
+
+                if (deleteEntry === 1) {
+                    return {
+                        status: 200,
+                        message: `Deleted ${medicineId} successfully`
+                    };
+                }
+
                 return {
-                    status: 200,
-                    message: `Deleted ${medicineId} successfully`
+                    status: 500,
+                    message: `Medicine ${medicineId} doesnt exist`
+                };
+            } else {
+                transaction_t1.rollback();
+                return {
+                    status: 500,
+                    message: `Medicine ${medicineId} doesnt exist`
                 };
             }
-
-            return {
-                status: 500,
-                message: `Medicine ${medicineId} doesnt exist`
-            };
         } catch (err: any) {
             console.log(err);
             return { status: 500, message: err.message };
